More Work on sonar findings
